Refactor Stats window view module
